разделили js-код на модули
